Finished marketplace loop (untested).  Refactored (cut in half), and readability fixed

+++ Server/src/Order.java
@@ -8,9 +8,9 @@ public class Order {
     private int orderID;
     private String datePlaced;
     private int assetID;
-    private float price;
+    private double price;
     private String orderType;
-    private int quantity;
+    private double quantity;
     private int userID;
     private String completed;
 
@@ -21,8 +21,8 @@ public class Order {
     public Order(ResultSet order){
         try {
             this.orderType = order.getString("OrderType");
-            this.price = order.getFloat("Price");
-            this.quantity = order.getInt("Quantity");
+            this.price = order.getDouble("Price");
+            this.quantity = order.getDouble("Quantity");
             this.assetID = order.getInt("AssetID");
             this.userID = order.getInt("UserID");
             this.orderID = order.getInt("OrderID");
@@ -36,7 +36,7 @@ public class Order {
     public Order(ResultSet order, String completed){
         try {
             this.orderType = order.getString("OrderType");
-            this.price = order.getFloat("Price");
+            this.price = order.getDouble("Price");
             this.quantity = order.getInt("Quantity");
             this.assetID = order.getInt("AssetID");
             this.userID = order.getInt("UserID");
@@ -62,15 +62,15 @@ public class Order {
 
     /**
      * Get the Price of the order
-     * @return Float of the price the order was placed for
+     * @return double of the price the order was placed for
      */
-    public float getPrice(){ return this.price; }
+    public double getPrice(){ return this.price; }
 
     /**
      * Get the Quantity of the order
      * @return Float of the quantity the order was placed for
      */
-    public int getQuantity(){ return this.quantity; }
+    public double getQuantity(){ return this.quantity; }
 
     /**
      * Get the Asset type that was placed in the order
@@ -86,15 +86,15 @@ public class Order {
 
     /**
      * Set the Price of the order
-     * @param price Float of the new price for the order
+     * @param price double of the new price for the order
      */
-    public void setPrice(float price){ this.price = price; }
+    public void setPrice(double price){ this.price = price; }
 
     /**
      * Set the Quantity of the order
-     * @param quantity Int of the new quantity for the order
+     * @param quantity double of the new quantity for the order
      */
-    public void setQuantity(int quantity){ this.quantity = quantity; }
+    public void setQuantity(double quantity){ this.quantity = quantity; }
 
     /**
      * Set the date the order was placed
